+++ src/main/java/javax/util/streamex/DoubleStreamEx.java
@@ -1208,6 +1208,36 @@ public class DoubleStreamEx implements DoubleStream {
                 i -> mapper.applyAsDouble(first[i], second[i]));
     }
 
+    /**
+     * Returns a stream consisting of the remaining elements of this stream
+     * after discarding the first {@code n} elements of the stream. If this
+     * stream contains fewer than {@code n} elements then an empty stream will
+     * be returned.
+     *
+     * <p>
+     * This is a stateful quasi-intermediate operation. Unlike
+     * {@link #skip(long)} it skips the first elements even if the stream is
+     * unordered. The main purpose of this method is to workaround the problem
+     * of skipping the first elements from non-sized source with further
+     * parallel processing and unordered terminal operation (such as
+     * {@link #forEach(DoubleConsumer)}). Also it behaves much better with infinite
+     * streams processed in parallel. For example,
+     * {@code DoubleStreamEx.iterate(0.0, i->i+1).skip(1).limit(100).parallel().toArray()}
+     * will likely to fail with {@code OutOfMemoryError}, but will work nicely
+     * if {@code skip} is replaced with {@code skipOrdered}.
+     *
+     * <p>
+     * For sequential streams this method behaves exactly like
+     * {@link #skip(long)}.
+     *
+     * @param n
+     *            the number of leading elements to skip
+     * @return the new stream
+     * @throws IllegalArgumentException
+     *             if {@code n} is negative
+     * @see #skip(long)
+     * @since 0.3.2
+     */
     public DoubleStreamEx skipOrdered(long n) {
         DoubleStream result = stream.isParallel() ? StreamSupport.doubleStream(
                 StreamSupport.doubleStream(stream.spliterator(), false).skip(n).spliterator(), true) : StreamSupport